Refactor: 몇몇 컴포넌트를 index.js로 이동
헤더, 푸터, 내비바 등을 import할 때 파일에서 직접 가져오는 게 아니라 index.js에서 가져오도록 함
