+++ src/components/Navbar/index.jsx
@@ -43,4 +43,4 @@ const Navbar = (props) => {
   );
 };
 
-export default Navbar;
+export { Navbar };
