+++ src/pages/FollowList/index.jsx
@@ -1,8 +1,7 @@
 import React from "react";
 
-import { Button, Img, List, Text } from "components";
+import { Button, Img, List, Text, MainHeader, Navbar } from "components";
 import FollowListFollowuserrow from "components/FollowListFollowuserrow";
-import FollowListNavbar from "components/FollowListNavbar";
 
 const FollowListPage = () => {
   const followListFollowuserrowPropList = [
@@ -15,35 +14,8 @@ const FollowListPage = () => {
   return (
     <>
       <div className="bg-white-A700 flex flex-col font-inter items-center justify-start mx-auto w-auto sm:w-full md:w-full">
-        <div className="bg-white-A700 flex flex-col items-center justify-start p-[26px] sm:px-5 w-full">
-          <div className="flex sm:flex-col flex-row md:gap-10 items-center justify-between max-w-[1180px] mx-auto md:px-5 w-full">
-            <div className="flex flex-col items-center justify-center px-4 w-auto">
-              <Img
-                className="h-[51px] md:h-auto object-cover w-[170px]"
-                src="images/img__51x170.png"
-                alt="FortyFive"
-              />
-            </div>
-            <div className="flex flex-row gap-2.5 items-start justify-start py-2.5 w-auto">
-              <Img
-                className="h-12 md:h-auto object-cover w-12"
-                src="images/img_bell111.png"
-                alt="bell111"
-              />
-              <Img
-                className="h-12 md:h-auto object-cover w-12"
-                src="images/img_search111.png"
-                alt="search111"
-              />
-              <Img
-                className="h-12 md:h-auto object-cover w-12"
-                src="images/img_menu11.png"
-                alt="menuEleven"
-              />
-            </div>
-          </div>
-        </div>
-        <FollowListNavbar className="flex flex-col items-center justify-start outline outline-blue_gray-100 p-2 w-full" />
+        <MainHeader className="bg-white-A700 flex flex-col items-center justify-start p-[25px] sm:px-5 w-full" />
+        <Navbar className="max-w-[1180px] mx-auto w-full" />
         <div className="border-b border-gray-300 border-solid flex flex-col gap-5 h-[603px] md:h-auto items-center justify-start max-w-[720px] pt-[30px] md:px-5 w-full">
           <div className="flex flex-row sm:gap-10 gap-20 items-start justify-center outline outline-gray-300 w-full">
             <div className="flex flex-col items-center justify-center pb-2.5 w-auto">
