+++ src/pages/LoginForm/index.jsx
@@ -1,8 +1,6 @@
 import React from "react";
 
-import { Button, CheckBox, Img, Input, Text } from "components";
-import FollowListNavbar from "components/FollowListNavbar";
-import Footer from "components/Footer";
+import { Button, CheckBox, Img, Input, Text, Footer, Navbar } from "components";
 import SearchResultHeader from "components/SearchResultHeader";
 
 const LoginFormPage = () => {
@@ -10,7 +8,7 @@ const LoginFormPage = () => {
     <>
       <div className="bg-white-A700 flex flex-col font-inter items-center justify-start mx-auto shadow-bs w-auto sm:w-full md:w-full">
         <SearchResultHeader className="bg-white-A700 flex flex-col items-center justify-start p-[26px] sm:px-5 w-full" />
-        <FollowListNavbar className="flex flex-col items-center justify-start outline outline-blue_gray-100 p-2 w-full" />
+        <Navbar className="flex flex-col items-center justify-start outline outline-blue_gray-100 p-2 w-full" />
         <div className="flex flex-col gap-5 h-[720px] md:h-auto items-center justify-start md:px-5 py-[120px] w-[720px] md:w-full">
           <div className="flex flex-row gap-2.5 items-center justify-center w-auto">
             <Img
