+++ src/pages/PwdModifyForm/index.jsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import { Img, Input, Text } from "components";
 import FollowListNavbar from "components/FollowListNavbar";
-import Footer from "components/Footer";
+import { Footer } from "components";
 import SearchResultHeader from "components/SearchResultHeader";
 
 const PwdModifyFormPage = () => {
