+++ src/pages/PwdResetForm/index.jsx
@@ -4,7 +4,7 @@ import { useNavigate } from "react-router-dom";
 
 import { Button, Img, Input, Text } from "components";
 import FollowListNavbar from "components/FollowListNavbar";
-import Footer from "components/Footer";
+import { Footer } from "components";
 import SearchResultHeader from "components/SearchResultHeader";
 
 const PwdResetFormPage = () => {
